Move some determinization code to separate modules DetUtils and ClassInterval
